requirements.txt
Flask
pyjwt
cryptography
Flask-Caching
